+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Added integration tests for `database/sql` extended API
+* Fixed extended API `internal/xsql/(*conn).GetTables(...)`, `internal/xsql/(*conn).GetAllTables(...)`
 * Fixed test for `database/sql` extended API
 * Extend `scheme.Client` interface with method `Database`
 * Fixed `sugar.RecursiveRemove` for remove full path